refactor: move_files uses a channel to report progress.
Implementing Progress as callback as move_files() did wasn't quite
correct in an async scenario. The problem was that to be Sync + Send,
the progress callback must block, using a blocking, non-tokio mutex,
which you're not supposed to do outside of spawn_blocking.

This change switches to using a channel to report progress and update
all callers of move_files.