fix: Move files from multiple dirs in the same queue.
This change avoids the strange output that happened when moving multiple
dirs, such as the number of files and bytes suddenly increasing after
copying has started.

This change waits until it's got the list of files from all dirs before
starting to copy, so progress information is correct and as expected for
realize-command.