feat!: Require 4-byte tag before RPC connection.
Before this change, whenever a client connected to the service port, the
RPC handler for RealizeService was started. This didn't allow any other
service to run on this port.

With this change, the client must specify the service they want to
connect to with a 4-byte tag at the beginning of the connection. REAL
for the realize service.