refactor: Hold one connector and acceptor instance in Networking.
Before this change, for each TcpStream a new connector and acceptor
created after splitting the private key to create a
RawPublicKeyResolver.

This change exposes RawPublicKeyResolver, so there's only one, and use
the opportunity to add convenience function to create it. It then use
that resolver in the Networking constructor to create one connector and
one acceptor, to be reused whenever necessary.