fix: Avoid many race conditions when watching a directory.
Before this change, real::history would go through the directory
hierarchy using WalkDir running on a blocking thread, which would report
what it finds through a channel.

This approach meant that we could go through the content of a directory
before watching it, instead of after, leaving the door open for file or
directories to be added between those two steps.

With this change, walking the directory is done in a more
straightforward way that guarantees that dirs are watched before their
content is read.