fix: Cache takes Arena into account, allocate inodes correctly.
Before this change, the unreal cache would take arenas as arguments but
completely ignore them. With this change, the cache tracks arenas and
stores them as special directories.

This change also fixes a bug in the inode allocation algorithm that
would not see inodes used by empty directories. To avoid this issue,
there's now a max_inode table.