refactoring: Transition unreal cache to indexer's notifications.
Before this change, the unreal cache expected old-style notification
sent by RealStore. This change switches to new-style notification sent
by the indexer and updates the code that sends and receives
notifications in Household.

Note that since indexers cannot be configured yet, the cache won't be
updated in production after this change. A follow-up to the setup is
required to get that working again.