fix: Use cache arenas when subscribing.
Before this chaneg, the test daemon_updates_cache would fail because the
client would not subscribe. The problem was that the client tried to
subscribe only to the arenas it had an index for, which made no sense.
This change fixes that.

This change also make daemon_updates_cache run on all OS, so it won't
remain silently broken again.