feat: The cache tracks hash version chain to choose the most recent.
This change modifies the logic used to choose the version returned by
the cache when more than one peer has a version.

Before this change, the version with the highest mtime was chosen
dynamically.

With this change, a decision is made about which version should be
returned and stored in a no-peer entry in addition to the peer entry.
That version tracks Replace(hash, old_hash), ignoring peers, to always
select the latest version according to the hash version chain.

This change also simplifies FileAvailability to have only one version
and one set of metadata, even when it is available from more than one
peer.

See the change in real.md for a discussion of the new algorithm.