refactor: Split cache into global and arena caches.
This change splits UnrealCacheBlocking into UnrealCacheBlocking and one
ArenaUnrealCacheBlocking per arena. This allows storing the arena cache
in the arena directory - or wherever the arena cache should use disk
space.

The goal of this refactoring is to prepare the way for adding a
blobstore to the cache, which stores file data which must go into the
arena dir.