refactor: Make cache required, arena root optional.
Before this change, cache was optional and arena root required. This
doesn't fit the expected usage and will make things hard when running
consensus, which requires a cache.

This change switches the configuration and Storage type to requiring a
cache. An arena root is, however, optional, but if one is set, an index
is created in the cache.