refactor(storage): Move away from using Arc for redb::Database.
redb::Database used to be stored in an Arc to allow different types to
share the same database. With the latest refactoring, the sharing is
done through an Arc<ArenaDatabase> instead, so ownership of
redb::Database can go back to being simple: Both the global cache and
ArenaDatabase own their redb::Database.