fix(storage): Report dirty paths to watcher after they're committed.
Before this change, dirty paths were reported to the watch channel
before they were even committed, which meant that the receiver could
very well run too early and not see the value reported by the channel.

This change solves the issue by allowing callbacks to be registered
after a transaction is committed successfully, to send out such
notifications.

It also uses the new after_commit callback in the index, to report
new history entries, which was previously done ad-hoc.