test: control_integration_test uses SetupHelper.
Before this changer control_integration_test started the daemon using
the command created by the crate realize-daemon. Unfortunately, rust
won't necessarily create this binary and there doesn't seem to be a way
to tell it to.

This change switches to using SetupHelper to setup the daemon and bind
to a unix path, which turned out to be simpler.