fix(storage): Detect empty files.
Before this change, empty files would be ignored by the RealWatcher,
because it did not react to Create(File), just Modify(Content) except
for hard links.

This change switches to creating a entry in the index for Create(File)
for empty files, even though we know that in most cases they'll quickly
be turned into non-empty files by a Modify(Content).

Before this change, the issue was hidden by catchup, which would in most
cases be able to catch the file created just after the watcher was
started. With this change, catchup is now run in tests only when it's
necessary, and it's *not* run in create_empty_file anymore.