feat: Add realize job action: move files from cache to FS.
With this change, the Engine detects file marked Own that need to be
moved out of the cache and into the filesystem. The job is then picked
up by Churten, which downloads and verifies the file before moving it.

Later on, if the file is updated remotely, it's updated on the index as
well.

One issue remains: if the file is removed remotely, it'll just stay on
the filesystem.