feat: Remove indexed file from FS when deleted remotely.
This change hook up the index with notifications so that it can detect
when a file in the index is outdated or removed.

If the file is outdated, update(notification) just updates the entry and
lets the Engine and Churten download the version into the cache and
update the file.

If the file is removed, update(notification) deletes the file right
away.

This moves the information about outdated index entries and the tracking
of it from the cache to the index itself.