refactor: realize to ArenaStorage, use only one transaction.
This change sets things up so that the index and the cache can use the
same write transaction for move_blob. The code is moved to ArenaStorage,
which has access to both types.

indexed_store::get_indexed_file is also the first function to move from
indexed_store.rs to index.rs