feat: Add Notification::Drop, a variation on Remove.
Notification::Drop allows removing a file from the index without it
being delete from other indexed.

This change defines the new notification and index history entry and
introduces a new method to the index to drop a file. (Unlike remove, drop
is not a result of fs modification, but rather some changes made during
churten.)

This will be used in the new "realize" job.