fix: Support reset in Churten.Subscriber in RPC
Before this change, sending to the subscriber would fail in
rpc::control::server whenever receiving from the channel would report a
lag. With this change, lag are supported and cause a reset, that is,
dropping all unhandled notifications in the queue and resetting the
active set of jobs - this way at least the client knows something is
wrong and can deal with it.