fix: Fix Storage::rsync when given offset > 0.
Before this change, Storage::rsync would only ever work on the beginning
of the file, because seek() was missing.

This change adds a seek() and unit tests for indexed_store::rsync in
storage, as well as more complete integration tests jobs::rsync in core,
so such issues will be detected in the future.