refactor: move tracking of client out of ConnectionManager
Before this change, ConnectionManager kept a map of Peer -> client to
call operations on.

This was awkward and limiting, as it meant that only the top
client (ConnnectedPeer) could be kept when we generally want a Store
client. It also made it impossible to do anything client-specific with
the client at initialization time.

This change moves the responsibility for tracking client instances to
the ConnectionHandler, now split into handler and tracker. When a
connection opens, ConnectionManager passes the top client to
register(peer, client) and the Houshold takes the store and subscribe to
it, then keep a handle on the store client. When a connection closes,
ConnectionManager calls unregister(peer) and the Houshold drops any
client it has kept.