feat: allow rate-limiting batch calls, client side
This change extends PeerConfig with a batch_rate_limit value. This value
is forwarded eventually to the Houshold which uses it to create
rate-limited client to use when making batch calls.

With this change, rsync and read calls must be flagged either
interactive (not rate-limited) or batch (may be rate-limited).