feat(core): jobs may report JobStatus::NoPeers
With this change, HouseholdOperationError::NoPeers that is returned
while a job is running is mapped to JobStatus::NoPeers, to tell the
Engine to retry next time a peer connects.

This change required moving away from anyhow::Error and use a custom
error for jobs, which can be matched by Churten.