test: attempt to reduce flakiness by waiting for specific hash
Before this change many tests wrote a file on one peer then waited for
it to appear in the cache of another peer. Unfortunately, due to
intermediate state sometimes being visible depending on notify, tests
could start too early, on an intermediate state, and fail.

This change makes tests wait for a specific version of a file to appear
in cache, corresponding to the final version, so they should be stabler.