fix: Avoid removing a file that an app will re-create right away.
When updating a file, some app delete the file and re-create it right
away. Before this change, this would result in a deletion followed by a
re-creation without history information.

This change handles the case specially, going through the hasher to
debounce deletion and checking before deleting.

This change also systematically uses fs_utils::metadata_no_symlink in
the indexer, to avoid issues with symlinks when checking for deletion.