feat(storage): global cache implements directories on its own
With this change, the global cache stops using the directory table and
logic of arena_cache, because it's becoming much too complicated for
this use case. The global cache just stores its own mini-hierarchy in
memory - just enough to deal with / in arena names and relies on the
arena caches for the rest.

This change replaces the table directory_table in the global db with
path_table.