refactor(arena_cache): define FileTableKey for the file table
This change introduces a Rust type to use as key for the file table, to
replace the tuple (inode, &str). This type has a slot for local version
information, though none is written yet.

The approach taken for this key is different from the one take for
values: instead of parsing and failing manually, this key has a
catch-all "Invalid" value that's returned when the byte representation
is invalid. To avoid having to parse the keys too often, byte-ordering
is used.