fix: always set disk_usage, account for inode disk usage
This change cleans up the way and time disk usage is set in blob
entries. It makes sure that disk usage is set at creation.

It also adds an arbitrary byte value to disk usage to account for inode
use and to avoid empty files looking like they're free when cleaning up.