refactor(tree): rework reference counting in tree
Before this change, add() would increase a reference count, reference
counting was otherwise left to the caller. This turned out to be
confusing and a source of errors.

This change introduces the rule that for a reference to a tree node,
there must be an entry in another table that represents that reference.
When the entry is created, a reference is added, when it is removed
the reference is removed.

This requires callers using the insert/remove methods in tree to insert
or remove these special entries/references, so it is not foolproof.