refactor(tree): update API to allow use of setup and setup_name
Before this change, the insert functions in WritableOpenTree would take
a path or a name and require that they be the one to setup the path or
name.

With the new method setup() and setup_name(), this isn't
necessary, since they check that no mapping with reference count 0
stays.

This change simplifies the API of the insert calls to rely on the caller
being able to call setup() if they want and pass in the inode they got
from it.

Reviewed-by: Stephane Zermatten <szermatt@gmx.net>