refactor(arena_cache): stop writing regular dir_table entries.
Before this change, the tree hierarchy was stored in dir_table and
tree_table. With this change, dir_table only contains the mtime of the
directories.

Before this change, dir_table regular entries were used to track tree
references, for reference counting. With this change, tree references
are tracked for file using the default entry in file_entry. They're not
tracked anymore for directories.

The above change means that directories might just disappear from view
once the last file in them is gone from the tree, even though their
mtime is still in the database. This will need to be fixed in follow-up
changes.