refactor(storage): give ownership of Tree to ArenaDatabase
With this change, ArenaDatabase takes ownership of tree so that
read_tree() and write_tree() can be called from transactions without
extra arguments.

This makes the setup of ArenaDatabase more complicated, which turned out
to be an issue in test. This change introduces "for_testing" and
"for_testing_single_arena" constructors for ArenaDatabase and ArenaCache
to hopefully keep the complexity in check, since these types are so
often created in tests.