refactor(engine): collect all db-handling code to DirtyPath
Before this change, Engine and DirtyPath were tightly connected in that
any change to the way the database worked would need to be applied to
both at the same time.

This change collects the code that interacts with the dirty tables into
DirtyPath and provide an API for Engine to call. It also adds tests
to test DirtyPath in isolation of Engine.

This paves the way towards making DirtyPath its own independent
lower-level module for other subsystems to call.