refactor(blob): decouple blob from file entry
With this change, BlobTableEntry contains the hash and size that used to
have to come each time from FileTableEntry. This makes blobs more self
contained, to the point where they can be used without necessarily going
through FileTableEntry. This also makes it possible to check blob
content against what's expected by the file.

This change also removes the blob_id from FileTableEntry, this just
duplicates information now that blob_id == inode. If we want to know
whether a blob is available, we can just query the blob table with the
inode.