refactor: move realize into storage, use inode
This change refactors StorageJobs realize and unrealize.

Before this change, Realize was an externally visible job, executed by
in the core layer. With this change, realize is executed within the
storage layer, after a cache has been downloaded by the core layer.

With this change, realize and unrealize use inodes instead of paths.