feat(dirty): give a reason when an inode is marked dirty
Giving a reason every time an inode is marked dirty, and logging that
reason, is meat to make it easier to understand why a specific inode is
processed.

This change also cleans up unnecessary TODOs in blob.rs, which only
needs to mark an inode dirty when it is verified.