refactor(index): get rid of RealIndex trait
This change replaces the RealIndex trait with a set of static functions
inside of index.rs that take an &Arc<ArenaDatabase>, as that's all the
state the index really needs.

These functions are themselves just convenience wrappers around the
transaction-based API.

This change also moves the index implementation out of arena_cache.rs
and back into index.rs.