refactor: turn HistoryReadOperation::history into an iterator
Before this change, HistoryReadOperation::history took a stream and
added the result to the stream. Not only was this awkward to use in
test, but it also caused issues if called (in test) from outside a
spawn_blocking()

This change switches the history method to return an iterator, which is
much more straightforward and convenient to use in tests and keeps the
blocking calls clearly inside spawn_blocking().