test(storage): work around macos spares file limitation
Before this change, the test created 1M files, wrote some data to it and
verified that the used size was the size of the written data, not the
file size - and that failed under macos.

On macos, files apparently have to be at least 32M to remain sparse once
data has been written to them.