test(storage): relax spare file disk usage test
Before this change, the test assumed that it can predict disk usage of
sparse files. While that works on Linux and some older version of macos,
the result were not consistent when run on the github macos virtual
machines with a newer version of macos.

With this change, the test require the files to take less than the
initial disk size, but doesn't specify how much less.