fix: realize creates index and sets mtime
Before this change, realize would move the file out of the blob store
and let the watcher eventually re-hash then index the file. This meant
that there was a time after realizing a file where the file looked like
it had to be realized again - so churten did just that.

This change also applies the mtime stored in the cache on the file and
stores that into the new index entry.