refactor!: reorganize arena paths
This change modifies the way arenas are configured. When before the
arena were given a database path, blob dir path and an optional root
path, they're now given a work directory and an optional data directory.

With this change the recommended layout is now to have the workdir be a
parent or sibling of datadir. This is required for overlayfs, which
needs a workdir of its out outside of datadir.

Example configurations:

[[arena]]
name = "example1"
workdir = "/store/example1"
datadir = "/store/example1/data"

In this example, the database is in /store/example1/arena.db and the
blobs are in /store/example1/blobs

[[arena]]
name = "example2"
workdir = "/store/.example2.realize"
datadir = "/store/example2"

With this example, realize-specific files are hidden, but still on the
same device as example2. The database is in
/store/.example2.realize/arena.db and the blobs are in
/store/.example2.realize/blobs.