feat!(cache): change the way removed peer versions are handled
Before this change, the cache tried hard to keep a version to serve as
long as any peer had one.

With this change, the cache instead considers that the file has been
deleted if it gets a Remove call with old_hash matching its current
hash.

This means that the file is not available in the cache right even, even
though peers might still have some version available.

BACKWARD INCOMPATIBILITY: This changes the way inodes are refcounted in
the database. Before this change, only the default entry mattered for
the inode refcount. With this change, peer versions matter as well, so
files might disappear unexpectedly when the new version run on against
an older database.