fix(cache): distinguish Notification::Drop from Remove.
Before this change, the cache would remove entries when other peers
reported them as dropped. This was not the intent; Drop just means that
the peer has decided not to keep or share this version anymore, not that
it is outdated.

This change only removes the peer entry when receiving a Drop
notification. It still removes the peer and the default entry, if it
matches, when receiving a Remove notification.