refactor: move file_availability to Blob::remote_availability
Before this change, file availability was separate from the blob. This
changes moves file_availability, renamed remote_availability into the
Blob, which matches very well with the way blobs are used in download.rs
and jobs.rs.

This is meant to allow Blobstore to redirect download to another
path/version than the file's in future changes.

This change also exposes the file version from file_metadata, for
testing, as tests very much relied on file_availability for that.