refactor(storage): extract ArenaCache::update
This change moves the implementation of ArenaCache::update into its own
function. This makes more sense, as update applies change to both the
cace and the index. Also it doesn't actually need anything but an
ArenaDatabase from ArenaCache.

This change still keeps ArenaCache::update and GlobalCache::update as
they are, though it is now possible to update::apply without having a
cache instance.