refactor!: rework HistoryTableEntry capnp definition
Before this change, HistoryTableEntry had a kind enum and fields that were
set or not depending on the kind. With this change, HistoryTableEntry
has big union with one entry for each kind and a struct for each one
that defines the fields to fill. This is clearer and better matches
the rust definition.