refactor: rename Inode to PathId
What used to be called Inode in the application doesn't match what's
normally called inode in file systems, since the inode is linked with a
path in a way that cannot change.

To avoid confusion, and since we need to work with real inodes, this
change renames Inode to PathId everywhere.