refactor: remove expect and resolve from GlobalCache and ArenaCache
Before this change, GlobalCache and ArenaCache exposed expect and
resolve from TreeReadOperations.

This was useful when the methods of these objects only took pathid, but
now, now so much, since all methods support a (Global)TreeLoc.

Additionally, having expect and resolve were on GlobalCache and
ArenaCache was confusing as it looked like they could be used to check
whether some file is in the cache, but only the tree level was checked
not the cache level. This use is best replaced by lookup(loc).