feat: add support for file renaming at the FUSE level
This change adds support for renaming files at the FUSE level in the
cache in a way that is later on reproduced by the owning peer.

To match FUSE expectations, the Inode of a renamed file follows the
name, which means that Inode and PathId numbers are sometimes swapped.

Renaming directory is not yet supported.