test: fix flaky test fuse::rename
In the rename scenario, there is a time where the renamed file cannot be
downloaded. This made the test flakey. This change adds retries when
reading the file.

It will be better to completely eliminate that time when the file is
unavailable, but that's a larger topic that'll take time.