feat(blob): add Blob::realize
This change adds a method to Blob to move the content of the blob to a
file in the datadir and returning the File underlying the Blob,
consuming it. It is useful when turning cached files into data files.

The method allows realizing any kind of blob, even partial ones. It's
the caller's responsibility to make sure the blob contains the data
that's needed.