feat(fuse): allow editing cached files
With this change, when a cached file is opened for write, it is first
downloaded and realized, then the realized file is made available trough
the FUSE interface.

This first implementation relies on the FUSE mountpoint being behind
mergerfs, which will redirect future calls to the files to the realized
file instead of hitting the cache, so the FUSE interface just needs to
allow editing the file for the lifetime of a file handle.