fix: drop outdated index entry
Before this change, when an index entry for a watched file was detected
by the engine, it would generate an Unrealize(), which would always be
later on abandoned since the index entry would not have the same hash as
the cached entry.

This change just drops the indexed entry in such a case instead of
unrealizing.