feat(fuse): specify how link, unlink and rename work on overwritten files
Cached files that have been overwritten are treated specially by the
FUSE interface, but just for a time as mergerfs quickly takes over, with
the file in datadir hiding the file in the cache. This change
nevertheless specifies what appens if link unlink or rename is applied
on a just overwritten file, for completeness and in case the FUSE
interface is accessed directly.