refactor!(cache): merge default and index layers
Before this change, entries for indexed (local) files where stored in
the index layer, then copied to the default layer. This was
unnecessarily complicated, since if a file was in the index layer, it
had to be in the default layer as well.

This change simplifies the situation by storing indexed files directly
in the default layer; there is no more index layer and no risk of the
two becoming inconsistent..