fix(storage): avoid timing issue due to preindex being delayed
Before this change, preindex and index in cache.rs would change the
database and otherwise trust that it made sense. With this change, index
checks that the file metadata (still) match and preindex takes metadata
from the file, which must exist.

This fix issues with the preindex call in watcher.rs being delayed
during awayt and adding to the database information that's out of date -
the file might even have been deleted at the time.