fix(watcher): read and check file metadata within a transaction
Before this change, file metadata was read once and forwarded to the
code that hashed the file. This resulted in inconsistencies, as the file
would often change between the time it was initially checked and the
time it was added to the database.

This change avoids inconsistencies by checking metadata within the
transaction and using these to schedule hashing and verifying, at the
end, that what was hashed didn't change.