fix: fix issue with TreeReadOperations::backtrack returning None
Before this change, TreeReadOperations::backtrack would return None when
it could not find a pathid as well as when the pathid it found was root.
This was in contradiction with its documentation, which stated that it
only returns None if the pathid was root. The end result is that we'd
get StorageError::IsADirectory for unknown pathids instead of
StorageError::NotFound. This changes fixes the inconsistency.