feat(storage): lookup detects new local files and dirs
Before this change, if a file or directory existed on the filesystem but
wasn't yet added to the cache by the watcher (usually because of a
inotify delay), that file or dir wouldn't be available in the filesystem
through lookup.

This change extends lookup to add entries to the cache as needed to
assign pathids to them, just like readdir does.