refactor(storage): add special files to cache, as local modification
This change adds the new FileEntryKind 'special' which is used for local
files that are not regular files, such as symlinks, fifo or devices.

These entries are only used as marker for existing local files. These
files are never meant to be indexed.

This change just introduces the new type; it doesn't yet make sure that
it is used properly in the watcher or when listing files.