refactor(path): introduce Path::root() to point to the root dir
Before this change, Path could only point to paths within the root. When
pointing to the root was necessary, Option<Path> was used instead. With
this change, path can point to the root as "".

This change also now also allows : inside of paths and distinguishes
paths with invalid portions (starting /, . or ..) from paths containing
invalide unicode.