fix(storage): metadata_no_symlink accept symlinks as last element
Before this change, if the last element of a path passed to
metadata_no_symlink was a symlink, the function would return NotFound.
With this change, it returns a Metadata with is_symlink() true. With
this changemetadata_no_symlink doesn't go *through* symlink dirs, but it
does return symlink files.

This is useful to be able to add symlinks to the cache as special local
file.