fix(watcher): rewrite watcher to support preindexed files
With this change the watcher preindexes all files and directories,
including special files and removes local files and directories from the
cache as they are deleted.

This is implemented as a rewrite of the watcher. The new implementation
behaves the same for every notification: it compares what's in the cache
with what's in the filesystem and updates the cache as needed, including
directories.

The scan portion of the watcher is now implemented as a series of events
sent to the main loop instead of a separate loop. This keep most of the
code common between the notification handling and the scanning logic.