refactor!(storage): store DiskUsageConfig in the database
This change modifies the settings table, storing a SettingsTableEntry
that includes uuid and DiskUsageConfig. DiskUsageConfig is, for now,
still set through configuration file, but the code is written in a way
that allows for setting it dynamically.

It also allows DiskUsageConfig.leave to be set without
DiskUsageConfig.max and replaces None, with an empty DiskUsageConfig.