fix!: sanitize the way global::Filesystem handles directories
Before this change, directories where stored in a confusing and
incorrect way in the database, then kept in a more straightforward
datastructure in memory.

This change rewrites PATH_TABLE to make it more like the memory
datastructure and to use Inode instead of PathId, since cross-arenas is
necessary.

This change fixes issues with intermediate paths in arena names having
the same name, like in "one/dir/arena" and "two/dir/arena", "dir" and
"arena" appear twice, but at different points in the hierarchy.

This change also fixes issues with the metadata returned by
global::Filesystem; readdir used to always return the same mtime for all
directories and report arena roots inconsistently with lookup.