refactor!: remove uses of Pathid, notable in ARENA_TABLE
The goal of this change is to get rid of the remaining uses of PathId as
a step towards removing it entirely.

A notable use is in ARENA_TABLE, to store the prefix. This change moves
away from using a PathId and instead stores the prefix as a u8 in a
capnp table entry, allowing for future backward-compatible changes.