refactor: add arena after creating Filesystem
This refactoring prepares the way for allowing the creation of new
arenas by refactoring the way FileSystem and Storage are created.

With this change, arena config is stored in the global database.
Filesystem can re-create the ArenaDatabase and ArenaFilesystem based on
that.

This change removes workdir from the configuration, but otherwise keeps
arenas in the config file. This is to be removed in a later change.