feat: don't preindex excluded files
Before this change, excluded files were not indexed, but they were
preindexed and would appear in the fuse filesystem.

With this change, excluded files are not preindexed by the watcher and
skipped when merging local files with cache data, so will not even
appear in the fuse filesystem.

This change also switches to using PathSet to store the exclusion set.