feat!: hardcode .realize subdirectory
Before this change, it was possible to specify both a data
directory (directory that stores local files) and a work
directory (directory that stores blob and arena database).

With this change, the work directory is always a subdirectory of the
data directory called .realize.

This simplifies setup and also makes it less likely to get into trouble
because the two directories are on different filesystems.