fix(core): subscribe to new arenas
Before this change, arenas were subscribed to at connection time, so if
a new arena was created after a connection was made, notification were
never sent.

With this change, the notification sender and receiver track new arenas
and subscribe to arenas when they are created, no matter the order in
which they're created.