test: remove ArenaDatabase::for_testing_no_blobs
Before this change, tests using ArenaDatabase::for_testing_to_blobs
would fail due to changes to ArenaDatabase that now requires the path to
be an directory, so the trick of passing in /dev/null in tests doesn't
work anymore.

This change removes the nonfunctional test helper and fixes the tests
that used it to create the necessary directory.