feat!: add a "default" mark type
Before this change, the default mark type was "watch", which could have
unintended consequences: if a local file was owned by another peer, it
would be automatically unrealized. This might come as a bad surprise.

This change adds a "default" mark, which works like watch minus
unrealize, meaning that a file in the cache will stay in the cache and a
local file will stay local.