feat: archive local files instead of deleting them directly
With this change, local files are archived instead of being deleted when
they are:
- deleted through FUSE (unlink)
- overwritten through a FUSE rename operation
- removed by selecting a remote version instead of the local version
- removed as instructed by a remote peer
- replaced with a remove version from another peer

This is meant to allow recovering and continuing to serve requests, but
this isn't part of this change yet.

Archives are blobs with special BlobIds included into the archived
queue. When disk space needs to be recovered, blobs in the archived
queue are deleted before blobs in the cached queue (previously known as
the working area).

The space taken up by archived blobs counts towards the evictable and
total disk space.